+++ frontend/client/analyze/analyze.css
