+++ frontend/client/viewer/components/pocket-list.css
