+++ frontend/client/viewer/viewer.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <link rel="stylesheet" href="./assets/css/jquery-ui.css">
+  <script type="text/javascript" src="./assets/js/jquery.min.js"></script>
+  <script type="text/javascript" src="./assets/js/jquery-ui.min.js"></script>
+  <script type="text/javascript" src="./assets/js/snap.svg-min.js"></script>
+  @require("../partials/head.html")
+</head>
+<body>
+@require("../partials/navigation.html")
+<noscript>
+  <strong>You need to enable JavaScript to run this app.</strong>
+</noscript>
+<div id="message">
+  <p class="text-center" id="message-text"></p>
+</div>
+<div id="analyze" class="display-none row">
+  <div id="application-rcsb" class="col-xs-12 col-md-8 col-xl-10"></div>
+  <div id="pocket-list-aside" class="col-xs-12 col-md-4 col-xl-2"></div>
+  <div id="application-molstar" class="col-xs-12 col-md-8 col-xl-10"></div>
+</div>
+@require("../partials/footer.html")
+</body>
+</html>
